#151: Refactor classes to use a parent class for not repeating code
