+++ src/Form/ImportVideoItemByMpxId.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 
 namespace Drupal\media_mpx\Form;
 
-use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\media_mpx\MpxLogger;
 use Drupal\media_mpx\Plugin\media\Source\Media;
@@ -18,36 +17,13 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *
  * @package Drupal\media_mpx\Form
  */
-class ImportVideoItemByMpxId extends FormBase {
+class ImportVideoItemByMpxId extends ImportUpdateVideoItem {
 
   /**
-   * The Update Video Item service.
-   *
-   * @var \Drupal\media_mpx\Service\UpdateVideoItem\UpdateVideoItem
-   */
-  private $updateVideoItem;
-
-  /**
-   * The media type repository.
-   *
-   * @var \Drupal\media_mpx\Repository\MpxMediaType
-   */
-  private $mpxTypeRepository;
-
-  /**
-   * The custom media mpx logger.
-   *
-   * @var \Drupal\media_mpx\MpxLogger
-   */
-  private $logger;
-
-  /**
-   * UpdateMediaItemForAccount constructor.
+   * {@inheritdoc}
    */
   public function __construct(UpdateVideoItem $updateVideoItem, MpxMediaType $mpxTypeRepository, MpxLogger $logger) {
-    $this->updateVideoItem = $updateVideoItem;
-    $this->mpxTypeRepository = $mpxTypeRepository;
-    $this->logger = $logger;
+    parent::__construct($updateVideoItem, $mpxTypeRepository, $logger);
   }
 
   /**
@@ -65,32 +41,14 @@ class ImportVideoItemByMpxId extends FormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
-    if (!$video_opts = $this->loadVideoTypeOptions()) {
-      $this->messenger()->addError($this->t('There has been an unexpected problem loading the form. Reload the page.'));
-      return [];
-    }
+    $form = parent::buildForm($form, $form_state);
 
-    $form['video_type'] = [
-      '#type' => 'select',
-      '#title' => $this->t('Video type'),
-      '#description' => $this->t('Choose the video type to import the video into.'),
-      '#options' => $video_opts,
-      '#default_value' => count($video_opts) === 1 ? array_keys($video_opts) : [],
-      '#required' => TRUE,
-    ];
     $form['mpx_id'] = [
       '#type' => 'textfield',
       '#title' => $this->t('ID'),
       '#placeholder' => $this->t('Type the ID of the mpx video you want to import.'),
       '#required' => FALSE,
     ];
-    $form['actions']['#type'] = 'actions';
-    $form['actions']['submit'] = [
-      '#type' => 'submit',
-      '#value' => $this->t('Import video'),
-      '#button_type' => 'primary',
-    ];
-
     return $form;
   }
 
@@ -121,30 +79,6 @@ class ImportVideoItemByMpxId extends FormBase {
       $this->messenger()->addError($this->t('There has been an unexpected problem importing the video. Check the logs for details.'));
       $this->logger->watchdogException($e);
     }
-  }
-
-  /**
-   * Returns the mpx Video Type options of the dropdown (prepared for form api).
-   *
-   * @return array
-   *   An array with options to show in the dropdown. The keys are the video
-   *   types, and the values are the video type label.
-   */
-  private function loadVideoTypeOptions(): array {
-    $video_opts = [];
-
-    try {
-      $video_types = $this->mpxTypeRepository->findAllTypes();
-
-      foreach ($video_types as $type) {
-        $video_opts[$type->id()] = $type->label();
-      }
-    }
-    catch (\Exception $e) {
-      $this->logger->watchdogException($e, 'Could not load mpx video type options.');
-    }
-
-    return $video_opts;
   }
 
   /**
