+++ src/Form/UpdateMediaItemForVideoType.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 
 namespace Drupal\media_mpx\Form;
 
-use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\media\MediaTypeInterface;
 use Drupal\media_mpx\DataObjectFactoryCreator;
@@ -24,28 +23,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *
  * @package Drupal\media_mpx\Form
  */
-class UpdateMediaItemForVideoType extends FormBase {
-
-  /**
-   * The update video service.
-   *
-   * @var \Drupal\media_mpx\Service\UpdateVideoItem\UpdateVideoItem
-   */
-  private $updateVideoItemService;
-
-  /**
-   * The mpx Media Type Repository.
-   *
-   * @var \Drupal\media_mpx\Repository\MpxMediaType
-   */
-  private $mpxTypeRepository;
-
-  /**
-   * The custom media mpx logger.
-   *
-   * @var \Drupal\media_mpx\MpxLogger
-   */
-  private $logger;
+class UpdateMediaItemForVideoType extends ImportUpdateVideoItem {
 
   /**
    * The Data Object Factory Creator.
@@ -67,9 +45,7 @@ class UpdateMediaItemForVideoType extends FormBase {
    *   The factory used to load a complete mpx object.
    */
   public function __construct(UpdateVideoItem $updateVideoItem, MpxMediaType $mpxTypeRepository, MpxLogger $logger, DataObjectFactoryCreator $dataObjectFactoryCreator) {
-    $this->updateVideoItemService = $updateVideoItem;
-    $this->mpxTypeRepository = $mpxTypeRepository;
-    $this->logger = $logger;
+    parent::__construct($updateVideoItem, $mpxTypeRepository, $logger);
     $this->dataObjectFactoryCreator = $dataObjectFactoryCreator;
   }
 
@@ -89,30 +65,12 @@ class UpdateMediaItemForVideoType extends FormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
-    if (!$video_opts = $this->loadVideoTypeOptions()) {
-      $this->messenger()->addError($this->t('There has been an unexpected problem loading the form. Reload the page.'));
-      return [];
-    }
-
-    $form['video_type'] = [
-      '#type' => 'select',
-      '#title' => $this->t('Video type'),
-      '#description' => $this->t('Choose the video type to import the video into.'),
-      '#options' => $video_opts,
-      '#default_value' => count($video_opts) === 1 ? array_keys($video_opts) : [],
-      '#required' => TRUE,
-    ];
+    $form = parent::buildForm($form, $form_state);
     $form['guid'] = [
       '#type' => 'textfield',
       '#title' => $this->t('GUID'),
       '#placeholder' => 'Type the GUID of the mpx video you want to import.',
       '#required' => TRUE,
-    ];
-    $form['actions']['#type'] = 'actions';
-    $form['actions']['submit'] = [
-      '#type' => 'submit',
-      '#value' => $this->t('Import video'),
-      '#button_type' => 'primary',
     ];
 
     return $form;
@@ -247,30 +205,6 @@ class UpdateMediaItemForVideoType extends FormBase {
 
     $results = $factory->select($query);
     return ($results->valid() ? $results->current() : NULL);
-  }
-
-  /**
-   * Returns the mpx Video Type options of the dropdown (prepared for form api).
-   *
-   * @return array
-   *   An array with options to show in the dropdown. The keys are the video
-   *   types, and the values are the video type label.
-   */
-  private function loadVideoTypeOptions(): array {
-    $video_opts = [];
-
-    try {
-      $video_types = $this->mpxTypeRepository->findAllTypes();
-
-      foreach ($video_types as $type) {
-        $video_opts[$type->id()] = $type->label();
-      }
-    }
-    catch (\Exception $e) {
-      $this->logger->watchdogException($e, 'Could not load mpx video type options.');
-    }
-
-    return $video_opts;
   }
 
 }
